requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
python-dotenv
gunicorn